Better attack handling
Characters now have option for a “defaultAttack” when no job is set.
Defaults to “Punch” afterwards.

Also new AttackResult struct for passing information back re: hit/miss
and critical.

+++ macguffin/Utility.swift
@@ -21,6 +21,6 @@ struct Utility {
     }
     
     struct Attack {
-        static let ChanceOfCriticalHit: Double = 0.05
+        static let ChanceOfCriticalHit: Float = 0.05
     }
 }
